requirements.txt
asgiref==3.8.1
certifi==2024.2.2
charset-normalizer==3.3.2
colorama==0.4.6
Django==5.0.5
django-jsonfield==1.4.1
djangorestframework==3.15.1
idna==3.7
iniconfig==2.0.0
jsonfield==3.1.0
packaging==24.0
pluggy==1.5.0
pytest==8.2.0
requests==2.31.0
six==1.16.0
sqlparse==0.5.0
tzdata==2024.1
urllib3==2.2.1
